samples/chat_langgraph/requirements.txt
openai==1.60.1
langchain==0.3.15
langchain-openai==0.3.2
langchain-community==0.3.15
opentelemetry-instrumentation-langchain==0.33.7
langgraph==0.2.73
pydantic==2.9.2
python-dotenv==1.0.1
azure-identity==1.19.0
streamlit==1.39.0
sqlalchemy==2.0.36
pyodbc==5.2.0
tiktoken==0.7.0
opentelemetry-instrumentation==0.48b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-distro[otlp]==0.48b0
opentelemetry-exporter-otlp
azure-monitor-opentelemetry-exporter==1.0.0b31
azure-monitor-opentelemetry==1.6.2
openinference-instrumentation-langchain==0.1.29
azure-search-documents==11.5.2
ipython==8.26.0
langgraph-cli[inmem]